backend/scripts/requirements.txt
feedparser
trafilatura
python-dotenv
qdrant-client
tqdm
requests
